refactor(util, view): separate validation logic and reorganize utility files
